Move Peripheral into peripheral module
Just some code cleanup, in preparation for possible changes in the
DMA module.
