+++ imxrt-hal/src/dma.rs
@@ -158,13 +158,13 @@ mod register;
 
 pub use buffer::{Buffer, Circular, CircularError, Linear, ReadHalf, WriteHalf};
 pub use element::Element;
-pub use memcpy::*;
+pub use memcpy::Memcpy;
+pub use peripheral::{helpers::*, Config, ConfigBuilder, Peripheral};
 
 use crate::{ccm, ral};
 use core::{
     fmt::{self, Debug, Display},
     mem,
-    sync::atomic::{compiler_fence, Ordering},
 };
 use register::{DMARegisters, MultiplexerRegisters, Static, DMA, MULTIPLEXER};
 
@@ -485,75 +485,6 @@ impl Channel {
     }
 }
 
-#[derive(Clone, Copy)]
-/// Configurations for defining DMA transfers
-///
-/// Use [`ConfigBuilder`](struct.ConfigBuilder.html) to create a DMA
-/// transfer configuration.
-pub struct Config {
-    interrupt_on_completion: bool,
-    interrupt_on_half: bool,
-    // Developer note: update `Config::from_channel` when
-    // adding more fields!
-}
-
-impl Config {
-    /// Returns a `Config` that represents the state of the supplied channel
-    fn from_channel(channel: &Channel) -> Self {
-        Config {
-            interrupt_on_completion: channel.is_interrupt_on_completion(),
-            interrupt_on_half: channel.is_interrupt_on_half(),
-        }
-    }
-}
-
-/// Builder for defining your DMA configuration
-///
-/// ```no_run
-/// use imxrt_hal::dma::ConfigBuilder;
-///
-/// let config = ConfigBuilder::new()
-///     .interrupt_on_completion(true)
-///     .build();
-/// ```
-pub struct ConfigBuilder(Config);
-
-impl ConfigBuilder {
-    /// Construct a builder, and begin defining a configuration
-    #[allow(clippy::new_without_default)] // Don't want to commit to a `Default` in the public interface
-    pub fn new() -> Self {
-        ConfigBuilder(Config {
-            interrupt_on_completion: false,
-            interrupt_on_half: false,
-        })
-    }
-
-    /// Specifies that this DMA channel will trigger an interrupt
-    /// when the transfer completes.
-    ///
-    /// The actual interrupt that will trigger depends on the supplied
-    /// channel. There are 15 interrupts for DMA channels, and each
-    /// interrupt supports two channels. You're responsible for managing
-    /// the interrupts, and for registering your handler for the correct
-    /// DMA channel.
-    pub fn interrupt_on_completion(mut self, interrupt_on_completion: bool) -> Self {
-        self.0.interrupt_on_completion = interrupt_on_completion;
-        self
-    }
-
-    /// Specifies that this DMA channel will trigger an interrupt when the transfer
-    /// is half complete
-    pub fn interrupt_on_half(mut self, interrupt_on_half: bool) -> Self {
-        self.0.interrupt_on_half = interrupt_on_half;
-        self
-    }
-
-    /// Complete configuration, and return a `Config` for a DMA transfer
-    pub fn build(self) -> Config {
-        self.0
-    }
-}
-
 /// An error when preparing a transfer
 #[derive(Debug)]
 #[non_exhaustive]
@@ -566,440 +497,6 @@ pub enum Error<P> {
     Peripheral(P),
     /// Error setting up the DMA transfer
     Setup(ErrorStatus),
-}
-
-impl<P> From<P> for Error<P> {
-    fn from(error: P) -> Self {
-        Error::Peripheral(error)
-    }
-}
-
-/// A DMA-capable peripheral
-///
-/// `Peripheral` wraps an object that can act as a source and / or destination
-/// for a DMA transfer. It provides an interface for scheduling transfers, and
-/// for knowing when transfers are complete.
-///
-/// When constructing a `Peripheral`, you may supply a configuration to trigger an interrupt when
-/// the DMA transfer completes. If you enable interrupts, you're responsible for registering the
-/// interrupt, and for clearing the interrupt. The `Peripheral` has methods for clearing interrupts
-/// due to transfer and receive DMA channels.
-///
-/// See the [module-level docs](index.html#example-full-duplex-spi-peripheral)
-/// for an example of how to create and use a peripheral.
-pub struct Peripheral<P, E, S, D = S> {
-    /// Channel used for outgoing data (from software, to external device)
-    tx_channel: Option<Channel>,
-    /// Channel used for incoming data (from external device, to software)
-    rx_channel: Option<Channel>,
-    /// The peripheral that is either providing the data, or accepting the data,
-    /// or both.
-    peripheral: P,
-    _element: core::marker::PhantomData<E>,
-    /// The buffer that used to send data in a DMA transfer
-    source_buffer: Option<S>,
-    /// The buffer that's used to receive data in a DMA transfer
-    destination_buffer: Option<D>,
-}
-
-impl<P, E, S, D> Peripheral<P, E, S, D> {
-    fn new(peripheral: P) -> Self {
-        Peripheral {
-            peripheral,
-            rx_channel: None,
-            tx_channel: None,
-            _element: core::marker::PhantomData,
-            source_buffer: None,
-            destination_buffer: None,
-        }
-    }
-}
-
-impl<P, E, S, D> Peripheral<P, E, S, D>
-where
-    P: peripheral::Source<E>,
-    E: Element,
-    D: buffer::Destination<E>,
-{
-    /// Wraps a peripheral that can act as the source of a DMA transfer
-    pub fn new_receive(source: P, channel: Channel, config: Config) -> Self {
-        let mut peripheral = Peripheral::new(source);
-        peripheral.init_receive(channel, config);
-        peripheral
-    }
-
-    fn init_receive(&mut self, mut channel: Channel, config: Config) {
-        channel.set_trigger_from_hardware(Some(P::SOURCE_REQUEST_SIGNAL));
-        // Safety: Source trait is only implemented on peripherals within
-        // this crate. We may study those implementations to show that the
-        // pointers point to valid memory.
-        unsafe {
-            channel.set_source(self.peripheral.source());
-        }
-        channel.set_interrupt_on_completion(config.interrupt_on_completion);
-        channel.set_interrupt_on_half(config.interrupt_on_half);
-        channel.set_disable_on_completion(true);
-        self.rx_channel = Some(channel);
-    }
-
-    /// Start a DMA transfer that transfers data from the peripheral into the supplied buffer
-    ///
-    /// A complete transfer is signaled by `is_receive_complete()`, and possibly an interrupt.
-    pub fn start_receive(&mut self, mut buffer: D) -> Result<(), (D, Error<P::Error>)> {
-        let rx_channel = self.rx_channel.as_mut().unwrap();
-        if rx_channel.is_enabled() {
-            return Err((buffer, Error::ScheduledTransfer));
-        }
-        let len = buffer.prepare_destination(rx_channel);
-        rx_channel.set_transfer_iterations(len as u16);
-        if let Err(error) = self.peripheral.enable_source() {
-            return Err((buffer, Error::Peripheral(error)));
-        }
-        compiler_fence(Ordering::Release);
-        rx_channel.set_enable(true);
-        if rx_channel.is_error() {
-            let es = ErrorStatus::new(rx_channel.error_status());
-            rx_channel.clear_error();
-            Err((buffer, Error::Setup(es)))
-        } else {
-            self.destination_buffer = Some(buffer);
-            Ok(())
-        }
-    }
-
-    /// Returns `true` if the receive is complete
-    pub fn is_receive_complete(&self) -> bool {
-        self.rx_channel.as_ref().unwrap().is_complete()
-    }
-
-    /// Clears the flag that indicates the DMA transfer is complete, and
-    /// disable the peripheral.
-    ///
-    /// Users are **required** to call this to disable the source. Otherwise,
-    /// the source may continue to generate DMA requests.
-    pub fn receive_complete(&mut self) -> Option<D> {
-        self.rx_channel.as_mut().unwrap().clear_complete();
-        self.peripheral.disable_source();
-        self.destination_buffer.take().and_then(|mut buffer| {
-            buffer.complete_destination();
-            Some(buffer)
-        })
-    }
-
-    /// Indicates if the receive channel has generated an interrupt
-    pub fn is_receive_interrupt(&self) -> bool {
-        self.rx_channel.as_ref().unwrap().is_interrupt()
-    }
-
-    /// Clears the interrupt flag on the receive channel
-    ///
-    /// Users are **required** to clear the interrupt flag, or the hardware
-    /// may continue to generate interrupts for the channel. This must be called
-    /// for completion interrupts and half-transfer interrupts.
-    pub fn receive_clear_interrupt(&mut self) {
-        self.rx_channel.as_mut().unwrap().clear_interrupt()
-    }
-
-    /// Cancel a receive transfer
-    pub fn receive_cancel(&mut self) -> Option<D> {
-        self.peripheral.disable_source();
-        let rx_channel = self.rx_channel.as_mut().unwrap();
-        while rx_channel.is_hardware_signaling() {
-            core::sync::atomic::spin_loop_hint();
-        }
-        rx_channel.set_enable(false);
-        compiler_fence(Ordering::Acquire);
-        self.destination_buffer.take()
-    }
-
-    /// Returns a copy of the config used to create the receive
-    /// peripheral
-    pub fn receive_config(&self) -> Config {
-        let rx_channel = self.rx_channel.as_ref().unwrap();
-        Config::from_channel(rx_channel)
-    }
-
-    /// Release the peripheral and the channel
-    ///
-    /// Users should ensure that any started transfer has completed. If the
-    /// `Peripheral` was constructed with [`new_bidirectional()`](struct.Peripheral.html#method.new_bidirectional),
-    /// callers should use [`bidirectional_release()`](struct.Peripheral.html#method.bidirectional_release);
-    /// otherwise, the transfer channel will be dropped when this method returns.
-    ///
-    /// To get a copy of the original config, use [`receive_config()`](struct.Peripheral.html#method.receive_config)
-    /// before releasing the object.
-    pub fn receive_release(mut self) -> (P, Channel) {
-        (self.peripheral, self.rx_channel.take().unwrap())
-    }
-
-    /// Indicates if the DMA controller is actively moving data for this DMA request
-    ///
-    /// It may not be moving data if
-    ///
-    /// - the transfer is complete, or there is no transfer
-    /// - the transfer is preempted
-    pub fn is_receive_active(&self) -> bool {
-        self.rx_channel.as_ref().unwrap().is_active()
-    }
-}
-
-impl<P, E, S> Peripheral<P, E, S, Circular<E>>
-where
-    P: peripheral::Source<E>,
-    E: Element,
-{
-    /// Returns the read half of the circular buffer that's being
-    /// used as a DMA transfer destination
-    ///
-    /// Returns `None` if there's no destination buffer, which may mean
-    /// that there's no active transfer.
-    pub fn read_half(&mut self) -> Option<ReadHalf<E>> {
-        self.destination_buffer.as_mut().map(ReadHalf::new)
-    }
-}
-
-/// Create a peripheral that can suppy `u8` data for DMA transfers
-pub fn receive_u8<P, B>(source: P, channel: Channel, config: Config) -> Peripheral<P, u8, B>
-where
-    P: peripheral::Source<u8>,
-    B: buffer::Destination<u8>,
-{
-    Peripheral::new_receive(source, channel, config)
-}
-
-/// Create a peripheral that can supply `u16` data for DMA transfers
-pub fn receive_u16<P, B>(source: P, channel: Channel, config: Config) -> Peripheral<P, u16, B>
-where
-    P: peripheral::Source<u16>,
-    B: buffer::Destination<u16>,
-{
-    Peripheral::new_receive(source, channel, config)
-}
-
-impl<P, E, S, D> Peripheral<P, E, S, D>
-where
-    P: peripheral::Destination<E>,
-    E: Element,
-    S: buffer::Source<E>,
-{
-    /// Wraps a peripheral that can act as the destination of a DMA transfer
-    pub fn new_transfer(destination: P, channel: Channel, config: Config) -> Self {
-        let mut peripheral = Peripheral::new(destination);
-        peripheral.init_transfer(channel, config);
-        peripheral
-    }
-
-    fn init_transfer(&mut self, mut channel: Channel, config: Config) {
-        channel.set_trigger_from_hardware(Some(P::DESTINATION_REQUEST_SIGNAL));
-        // Safety: Destination trait is only implemented on peripherals within
-        // this crate. We may study those implementations to show that the pointers
-        // point to valid memory.
-        unsafe {
-            channel.set_destination(self.peripheral.destination());
-        }
-        channel.set_interrupt_on_completion(config.interrupt_on_completion);
-        channel.set_interrupt_on_half(config.interrupt_on_half);
-        channel.set_disable_on_completion(true);
-        self.tx_channel = Some(channel);
-    }
-
-    /// Start a DMA transfer that transfers data from the supplied buffer to the peripheral
-    ///
-    /// A complete transfer is signaled by `is_transfer_complete()`, and possibly an interrupt.
-    pub fn start_transfer(&mut self, mut buffer: S) -> Result<(), (S, Error<P::Error>)> {
-        let tx_channel = self.tx_channel.as_mut().unwrap();
-        if tx_channel.is_enabled() {
-            return Err((buffer, Error::ScheduledTransfer));
-        }
-        let len = buffer.prepare_source(tx_channel);
-        tx_channel.set_transfer_iterations(len as u16);
-        if let Err(error) = self.peripheral.enable_destination() {
-            return Err((buffer, Error::Peripheral(error)));
-        }
-        compiler_fence(Ordering::Release);
-        tx_channel.set_enable(true);
-        if tx_channel.is_error() {
-            let es = ErrorStatus::new(tx_channel.error_status());
-            tx_channel.clear_error();
-            Err((buffer, Error::Setup(es)))
-        } else {
-            self.source_buffer = Some(buffer);
-            Ok(())
-        }
-    }
-
-    /// Returns `true` if the transfer is complete
-    pub fn is_transfer_complete(&self) -> bool {
-        self.tx_channel.as_ref().unwrap().is_complete()
-    }
-
-    /// Clears the flag that indicates the DMA transfer is complete, and
-    /// disable the peripheral.
-    ///
-    /// Users are **required** to call this to disable the source. Otherwise,
-    /// the source may continue to generate DMA requests.
-    pub fn transfer_complete(&mut self) -> Option<S> {
-        self.tx_channel.as_mut().unwrap().clear_complete();
-        self.peripheral.disable_destination();
-        self.source_buffer.take().and_then(|mut buffer| {
-            buffer.complete_source();
-            Some(buffer)
-        })
-    }
-
-    /// Indicates if the transfer channel has generated an interrupt
-    pub fn is_transfer_interrupt(&self) -> bool {
-        self.tx_channel.as_ref().unwrap().is_interrupt()
-    }
-
-    /// Clears the interrupt flag on the transfer channel
-    ///
-    /// Users are **required** to clear the interrupt flag, or the hardware
-    /// may continue to generate interrupts for the channel. This must be called
-    /// for completion interrupts and half-transfer interrupts.
-    pub fn transfer_clear_interrupt(&mut self) {
-        self.tx_channel.as_mut().unwrap().clear_interrupt()
-    }
-
-    /// Cancel a transfer that sends data to the peripheral
-    pub fn transfer_cancel(&mut self) -> Option<S> {
-        self.peripheral.disable_destination();
-        let tx_channel = self.tx_channel.as_mut().unwrap();
-        while tx_channel.is_hardware_signaling() {
-            core::sync::atomic::spin_loop_hint();
-        }
-        tx_channel.set_enable(false);
-        compiler_fence(Ordering::Acquire);
-        self.source_buffer.take()
-    }
-
-    /// Returns a copy of the transfer config supplied during
-    /// construction
-    pub fn transfer_config(&self) -> Config {
-        let tx_channel = self.tx_channel.as_ref().unwrap();
-        Config::from_channel(tx_channel)
-    }
-
-    /// Release the peripheral and the channel
-    ///
-    /// Users should ensure that any started transfer has completed. If the
-    /// `Peripheral` was constructed with [`new_bidirectional()`](struct.Peripheral.html#method.new_bidirectional),
-    /// callers should use [`bidirectional_release()`](struct.Peripheral.html#method.bidirectional_release);
-    /// otherwise, the receiver channel will be dropped when this method returns.
-    ///
-    /// To get a copy of the original config, use [`transfer_config()`](struct.Peripheral.html#method.transfer_config)
-    /// before releasing the object.
-    pub fn transfer_release(mut self) -> (P, Channel) {
-        (self.peripheral, self.tx_channel.take().unwrap())
-    }
-
-    /// Indicates if the DMA controller is actively moving data for this DMA request
-    ///
-    /// It may not be moving data if
-    ///
-    /// - the transfer is complete, or there is no transfer
-    /// - the transfer is preempted
-    pub fn is_transfer_active(&self) -> bool {
-        self.rx_channel.as_ref().unwrap().is_active()
-    }
-}
-
-impl<P, E, D> Peripheral<P, E, Circular<E>, D>
-where
-    P: peripheral::Destination<E>,
-    E: Element,
-{
-    /// Returns the write half of the circular buffer that's being
-    /// used as a DMA transfer source.
-    ///
-    /// Returns `None` if there's no source buffer, which may mean
-    /// that there's no active transfer.
-    ///
-    /// You may use the [`WriteHalf`](struct.WriteHalf.html) to prepare
-    /// data for another DMA transfer.
-    pub fn write_half(&mut self) -> Option<WriteHalf<E>> {
-        self.source_buffer.as_mut().map(WriteHalf::new)
-    }
-}
-
-/// Create a peripheral that can accept `u8` data from DMA transfers
-pub fn transfer_u8<P, B>(destination: P, channel: Channel, config: Config) -> Peripheral<P, u8, B>
-where
-    P: peripheral::Destination<u8>,
-    B: buffer::Source<u8>,
-{
-    Peripheral::new_transfer(destination, channel, config)
-}
-
-/// Create a peripheral that can accept `u16` data from DMA transfers
-pub fn transfer_u16<P, B>(destination: P, channel: Channel, config: Config) -> Peripheral<P, u16, B>
-where
-    P: peripheral::Destination<u16>,
-    B: buffer::Source<u16>,
-{
-    Peripheral::new_transfer(destination, channel, config)
-}
-
-impl<P, E, S, D> Peripheral<P, E, S, D>
-where
-    P: peripheral::Source<E> + peripheral::Destination<E>,
-    E: Element,
-    S: buffer::Source<E>,
-    D: buffer::Destination<E>,
-{
-    /// Wraps a peripheral that can act as both the source and destination of a DMA transfer
-    pub fn new_bidirectional(peripheral: P, tx: (Channel, Config), rx: (Channel, Config)) -> Self {
-        let mut peripheral = Peripheral::new(peripheral);
-        peripheral.init_receive(rx.0, rx.1);
-        peripheral.init_transfer(tx.0, tx.1);
-        peripheral
-    }
-
-    /// Release the peripheral and both channels (transfer channel, release channel)
-    ///
-    /// Users should ensure that any active transfers are complete before releasing the
-    /// peripheral.
-    pub fn bidirectional_release(mut self) -> (P, (Channel, Channel)) {
-        (
-            self.peripheral,
-            (
-                self.tx_channel.take().unwrap(),
-                self.rx_channel.take().unwrap(),
-            ),
-        )
-    }
-}
-
-/// Create a peripheral that can accept `u8` data from DMA transfers, and can
-/// source `u8` data for DMA transfers
-pub fn bidirectional_u8<P, S, D>(
-    peripheral: P,
-    tx: (Channel, Config),
-    rx: (Channel, Config),
-) -> Peripheral<P, u8, S, D>
-where
-    P: peripheral::Source<u8, Error = <P as peripheral::Destination<u8>>::Error>
-        + peripheral::Destination<u8>,
-    S: buffer::Source<u8>,
-    D: buffer::Destination<u8>,
-{
-    Peripheral::new_bidirectional(peripheral, tx, rx)
-}
-
-/// Create a peripheral that can accept `u16` data from DMA transfers, and can
-/// source `u16` data for DMA transfers
-pub fn bidirectional_u16<P, S, D>(
-    peripheral: P,
-    tx: (Channel, Config),
-    rx: (Channel, Config),
-) -> Peripheral<P, u16, S, D>
-where
-    P: peripheral::Source<u16, Error = <P as peripheral::Destination<u16>>::Error>
-        + peripheral::Destination<u16>,
-    S: buffer::Source<u16>,
-    D: buffer::Destination<u16>,
-{
-    Peripheral::new_bidirectional(peripheral, tx, rx)
 }
 
 /// Unclocked, uninitialized DMA channels
